SeafileProvider: do file download in threadpool
This is a workaround. Many users of GET_DOCUMENT/OPEN_DOCUMENT
don't realize that opening a file might cause network access and thus do
them in the ui thread. This results in a NetworkOnMainThreadException.

Therefore do the download in another thread and block until the download
is complete. This is also the behaviour of Google Drive's DocumentProvider
implementation.

This fixes NetworkOnMainThreadExceptions in

* Android: Settings->Security->Credential storage->Install from storage
* OpenKeychain: import certificate
* OpenVPN for Android: import settings file